move tables_information.py
